refactor: Ball 도메인, 패키지 분리
