+++ src/main/java/lotto/domain/ball/Balls.java
@@ -1,4 +1,4 @@
-package lotto.domain.ticket;
+package lotto.domain.ball;
 
 import java.util.HashMap;
 import java.util.Map;
