+++ src/main/java/lotto/domain/ticket/Tickets.java
@@ -5,6 +5,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
+import lotto.domain.ball.Ball;
 import lotto.domain.rank.Rank;
 import lotto.domain.ticket.generator.TicketGenerator;
