+++ src/main/java/lotto/domain/ticket/generator/RandomTicketGenerator.java
@@ -4,8 +4,8 @@ import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import lotto.domain.ball.BallNumberRange;
 import lotto.domain.ticket.Ticket;
-import lotto.domain.ticket.condition.BallNumberRange;
 import lotto.domain.ticket.condition.TicketSize;
 
 public class RandomTicketGenerator implements TicketGenerator {
