+++ src/main/java/lotto/service/LottoService.java
@@ -2,11 +2,11 @@ package lotto.service;
 
 import java.util.List;
 
+import lotto.domain.ball.Ball;
+import lotto.domain.ball.Balls;
 import lotto.domain.money.Money;
 import lotto.domain.rank.Rank;
 import lotto.domain.ticket.Analysis;
-import lotto.domain.ticket.Ball;
-import lotto.domain.ticket.Balls;
 import lotto.domain.ticket.Ticket;
 import lotto.domain.ticket.Tickets;
 import lotto.domain.ticket.WinningTicket;
